test(relay): serialize env-touching tests on one lock
The per-group mutexes only serialized tests that mutate the same variable
group. clap reads the MOQ_* vars while parsing, so a test clearing its own
group could still race a sibling group's `Config::parse_and_merge` reading
the environment, which is exactly the data race `set_var`/`remove_var` are
unsafe for. The two config-merge tests in cluster.rs took no lock at all.

Replace the twelve per-group mutexes with one ENV_LOCK, and fold it into
EnvGuard so the lock cannot be forgotten: the guard holds it from the
snapshot through the restore in Drop. `EnvGuard::lock()` covers the tests
that only read the environment via clap.
